requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
django-dramatiq==0.9.1    # via -r requirements.in
django==3.1               # via -r requirements.in, django-dramatiq, djangorestframework
djangorestframework==3.11.1  # via -r requirements.in
dramatiq==1.9.0           # via django-dramatiq
factory-boy==3.0.1        # via -r requirements.in
faker==4.1.2              # via factory-boy
idna==2.10                # via requests
lxml==4.5.2               # via -r requirements.in
prometheus-client==0.8.0  # via dramatiq
psycopg2-binary==2.8.5    # via -r requirements.in
python-dateutil==2.8.1    # via faker
python-dotenv==0.14.0     # via -r requirements.in
pytz==2020.1              # via django
redis==3.5.3              # via -r requirements.in
requests==2.24.0          # via -r requirements.in
six==1.15.0               # via python-dateutil
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
urllib3==1.25.10          # via requests
